# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API SmartQuote",
    "description": "API para gerenciamento do SmartQuote.",
    "version": "1.0.0"
  },
  "host": "https://testsmart-24vt.onrender.com",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Listar todos os usuários",
        "description": "Retorna uma lista de todos os usuários cadastrados.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/cotacoes": {
      "get": {
        "summary": "Listar todas as cotações",
        "description": "Retorna uma lista de todas as cotações com informações de prompt e produto.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de cotações retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar uma nova cotação",
        "description": "Cria uma nova cotação no sistema.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados da cotação para criação",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "prompt_id": {
                  "type": "integer",
                  "example": 1
                },
                "produto_id": {
                  "type": "integer",
                  "example": 5
                },
                "preco": {
                  "type": "number",
                  "example": 1200.50
                },
                "observacao": {
                  "type": "string",
                  "example": "Entrega em 7 dias"
                }
              },
              "required": [
                "prompt_id",
                "produto_id",
                "preco"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Cotação criada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/produtos": {
      "get": {
        "summary": "Listar todos os produtos",
        "description": "Retorna uma lista de todos os produtos cadastrados.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar um novo produto",
        "description": "Cria um novo produto no sistema.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados do produto para criação",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Notebook Dell"
                },
                "preco": {
                  "type": "number",
                  "example": 3500
                },
                "estoque": {
                  "type": "integer",
                  "example": 10
                },
                "descricao": {
                  "type": "string",
                  "example": "Notebook com 16GB RAM"
                }
              },
              "required": [
                "nome",
                "preco",
                "estoque"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Produto criado com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/fornecedores": {
      "get": {
        "summary": "Listar todos os fornecedores",
        "description": "Retorna uma lista de todos os fornecedores cadastrados.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de fornecedores retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar um novo fornecedor",
        "description": "Cria um novo fornecedor no sistema.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados do fornecedor para criação",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FornecedorCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Fornecedor criado com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "summary": "Listar todas as notificações",
        "description": "Retorna uma lista de todas as notificações.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de notificações retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar uma nova notificação",
        "description": "Cria uma nova notificação no sistema.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados da notificação para criação",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 1
                },
                "user_id": {
                  "type": "string",
                  "example": "uuid-do-usuario"
                },
                "title": {
                  "type": "string",
                  "example": "Nova Cotação Recebida"
                },
                "body": {
                  "type": "string",
                  "example": "Uma nova cotação foi adicionada ao sistema."
                }
              },
              "required": [
                "user_id",
                "title",
                "body"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Notificação criada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/sistema": {
      "get": {
        "summary": "Buscar configurações do sistema",
        "description": "Retorna as configurações atuais do sistema (nome, idioma, fuso horário, moeda, backup).",
        "tags": ["Sistema"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Configurações encontradas",
            "schema": {
              "$ref": "#/definitions/SistemaResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar ou atualizar configurações do sistema",
        "description": "Salva as configurações do sistema (nome, idioma, fuso horário, moeda, backup).",
        "tags": ["Sistema"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "sistema",
            "description": "Dados das configurações do sistema",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SistemaCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configurações salvas com sucesso",
            "schema": {
              "$ref": "#/definitions/SistemaResponse"
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "patch": {
        "summary": "Atualizar parcialmente configurações do sistema",
        "description": "Atualiza apenas os campos especificados das configurações do sistema.",
        "tags": ["Sistema"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "sistema",
            "description": "Campos a serem atualizados",
            "schema": {
              "$ref": "#/definitions/SistemaUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configurações atualizadas com sucesso",
            "schema": {
              "$ref": "#/definitions/SistemaResponse"
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Configurações não encontradas"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    }
  },
  "definitions": {
...
    "FornecedorCreate": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Fornecedor X"
        },
        "email": {
          "type": "string",
          "format": "email",
          "example": "contato@fornecedorx.com"
        },
        "telefone": {
          "type": "string",
          "example": "+55 11 99999-9999"
        }
      },
      "required": [
        "nome",
        "email"
      ]
    },
...
    "SistemaCreate": {
      "type": "object",
      "properties": {
        "nome_empresa": {
          "type": "string",
          "example": "SmartQuote Ltda"
        },
        "idioma": {
          "type": "string",
          "example": "pt-BR"
        },
        "fuso_horario": {
          "type": "string",
          "example": "America/Sao_Paulo"
        },
        "moeda": {
          "type": "string",
          "example": "BRL"
        },
        "backup": {
          "type": "string",
          "example": "diario"
        },
        "manutencao": {
          "type": "boolean",
          "example": false
        },
        "tempo_de_sessao": {
          "type": "integer",
          "example": 480
        },
        "politica_senha": {
          "type": "string",
          "enum": ["forte", "medio"],
          "example": "forte"
        },
        "log_auditoria": {
          "type": "boolean",
          "example": true
        },
        "ip_permitidos": {
          "type": "string",
          "example": "192.168.1.0/24,10.0.0.0/8"
        }
      },
      "required": [
        "nome_empresa",
        "idioma",
        "fuso_horario",
        "moeda",
        "backup",
        "manutencao",
        "tempo_de_sessao",
        "politica_senha",
        "log_auditoria",
        "ip_permitidos"
      ]
    },
    "SistemaUpdate": {
      "type": "object",
      "properties": {
        "nome_empresa": {
          "type": "string",
          "example": "SmartQuote Pro Ltda"
        },
        "idioma": {
          "type": "string",
          "example": "en-US"
        },
        "fuso_horario": {
          "type": "string",
          "example": "America/New_York"
        },
        "moeda": {
          "type": "string",
          "example": "USD"
        },
        "backup": {
          "type": "string",
          "example": "semanal"
        },
        "manutencao": {
          "type": "boolean",
          "example": true
        },
        "tempo_de_sessao": {
          "type": "integer",
          "example": 360
        },
        "politica_senha": {
          "type": "string",
          "enum": ["forte", "medio"],
          "example": "medio"
        },
        "log_auditoria": {
          "type": "boolean",
          "example": false
        },
        "ip_permitidos": {
          "type": "string",
          "example": "10.0.0.0/8"
        }
      }
    },
    "SistemaResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "Configurações do sistema encontradas"
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "nome_empresa": {
              "type": "string",
              "example": "SmartQuote Ltda"
            },
            "idioma": {
              "type": "string",
              "example": "pt-BR"
            },
            "fuso_horario": {
              "type": "string",
              "example": "America/Sao_Paulo"
            },
            "moeda": {
              "type": "string",
              "example": "BRL"
            },
            "backup": {
              "type": "string",
              "example": "diario"
            },
            "manutencao": {
              "type": "boolean",
              "example": false
            },
            "tempo_de_sessao": {
              "type": "integer",
              "example": 480
            },
            "politica_senha": {
              "type": "string",
              "enum": ["forte", "medio"],
              "example": "forte"
            },
            "log_auditoria": {
              "type": "boolean",
              "example": true
            },
            "ip_permitidos": {
              "type": "string",
              "example": "192.168.1.0/24,10.0.0.0/8"
            },
            "criado_em": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-29T10:00:00.000Z"
            },
            "atualizado_em": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-29T10:00:00.000Z"
            }
          }
        }
      }
    }
  }
}